+++ Cargo.toml
@@ -5,4 +5,4 @@ resolver = "2"
 [workspace.dependencies]
 borsh = "1.5.1"
 solana-program = { git = "https://github.com/anza-xyz/agave" }
-groth16-solana = { path = "verifier" }
+sp1-solana = { path = "verifier" }
