+++ example/README.md
@@ -1,32 +1,40 @@
 # Example
 
-This example demonstrates how to use the `groth16-solana` crate to verify a proof generated by the `SP1` library on Solana.
+This example demonstrates how to use the `sp1-solana` crate to verify a proof generated by SP1 on solana.
 
 ## Overview: Script
 
-This script will load a serialized [`SP1ProofWithPublicValues`](https://docs.rs/sp1-sdk/2.0.0/sp1_sdk/proof/struct.SP1ProofWithPublicValues.html)
+This script will load a serialized [`SP1ProofWithPublicValues`]
+(https://docs.rs/sp1-sdk/2.0.0/sp1_sdk/proof/struct.SP1ProofWithPublicValues.html)
 from the pre-generated proof [`fibonacci_proof.bin`](./proofs/fibonacci_proof.bin). This is a SP1 Groth16 proof that
 proves that the 20th fibonacci number is 6765.
 
 The loaded `SP1ProofWithPublicValues` will be serialized into a `SP1ProofFixture`, and written to
-[`proof-fixtures/fibonacci_fixture.bin`](./proof-fixtures/fibonacci_fixture.bin). The `SP1ProofFixture` can easily be borsh serialized and 
-deserialized, and is used to verify the proof. 
+[`proof-fixtures/fibonacci_fixture.bin`](./proof-fixtures/fibonacci_fixture.bin). The `SP1ProofFixture`
+can easily be borsh serialized and deserialized, and can be verified on chain.
 
 ### Running the script
 
-Run the following commands. 
+To load the pregenerated proof and verify it, run the following commands. 
 
 ```shell
 cd script
 RUST_LOG=info cargo run --release
 ```
 
+To generate a fresh proof from the program in `sp1-program`, run the following commands. 
+
+```shell
+cd script
+RUST_LOG=info cargo run --release -- --prove
+```
+
 ## Overview: Solana Program
 
-The code in [`solana/program`](./solana/program) is a simple Solana program that verifies a `SP1ProofFixture` using the `groth16-solana` crate.
-It costs roughly 280,000 compute units. 
+The code in [`solana/program`](./solana/program) is a simple solana contract that verifies a 
+`SP1ProofFixture` using the `sp1-solana` crate. It costs roughly 280,000 compute units. 
 
-Here is an excerpt from the solana program that demonstrates how to verify a `SP1ProofFixture` on chain. 
+Here is an excerpt from the solana contract that demonstrates how to verify a `SP1ProofFixture` on chain. 
 
 ```rust
 pub fn process_instruction(
@@ -37,8 +45,8 @@ pub fn process_instruction(
     // Deserialize the fixture from the instruction data.
     let fixture = SP1ProofFixture::try_from_slice(instruction_data).unwrap();
 
-    // Get the SP1 Groth16 verification key from the `groth16-solana` crate.
-    let vk = groth16_solana::GROTH16_VK_BYTES;
+    // Get the SP1 Groth16 verification key from the `sp1-solana` crate.
+    let vk = sp1_solana::GROTH16_VK_BYTES;
 
     // Verify the proof.
     let result = verify_proof_fixture(&fixture, &vk);
