+++ example/script/src/main.rs
@@ -1,5 +1,20 @@
-use groth16_solana::{verify_proof_fixture, SP1ProofFixture, GROTH16_VK_BYTES};
-use sp1_sdk::{utils, SP1ProofWithPublicValues};
+use clap::Parser;
+use sp1_sdk::{utils, ProverClient, SP1ProofWithPublicValues, SP1Stdin};
+use sp1_solana::{verify_proof_fixture, SP1ProofFixture, GROTH16_VK_BYTES};
+
+#[derive(clap::Parser)]
+#[command(name = "zkVM Proof Generator")]
+struct Cli {
+    #[arg(
+        long,
+        value_name = "prove",
+        default_value = "false",
+        help = "Specifies whether to generate a proof for the program."
+    )]
+    prove: bool,
+}
+
+const ELF: &[u8] = include_bytes!("../../sp1-program/elf/riscv32im-succinct-zkvm-elf");
 
 fn main() {
     // Setup logging for the application.
@@ -7,6 +22,30 @@ fn main() {
 
     // Where to save / load the proof from.
     let proof_file = "../../proofs/fibonacci_proof.bin";
+
+    // Parse command line arguments.
+    let args = Cli::parse();
+
+    // Only generate a proof if the prove flag is set.
+    if args.prove {
+        // Initialize the prover client
+        let client = ProverClient::new();
+        let (pk, _) = client.setup(ELF);
+
+        // Compute the 20th fibonacci number.
+        let mut stdin = SP1Stdin::new();
+        stdin.write(&20u32);
+
+        // Generate a proof for the fibonacci program.
+        let proof = client
+            .prove(&pk, stdin)
+            .groth16()
+            .run()
+            .expect("Groth16 proof generation failed");
+
+        // Save the generated proof to `proof_file`.
+        proof.save(&proof_file).unwrap();
+    }
 
     // Load the proof from the file, and convert it to a fixture.
     let sp1_proof_with_public_values = SP1ProofWithPublicValues::load(&proof_file).unwrap();
