+++ example/solana/program/src/lib.rs
@@ -1,6 +1,6 @@
 use borsh::BorshDeserialize;
-use groth16_solana::{verify_proof_fixture, SP1ProofFixture};
 use solana_program::{account_info::AccountInfo, entrypoint::ProgramResult, msg, pubkey::Pubkey};
+use sp1_solana::{verify_proof_fixture, SP1ProofFixture};
 
 #[cfg(not(feature = "no-entrypoint"))]
 use solana_program::entrypoint;
@@ -17,7 +17,7 @@ pub fn process_instruction(
     let fixture = SP1ProofFixture::try_from_slice(instruction_data).unwrap();
 
     // Get the SP1 Groth16 verification key from the `groth16-solana` crate.
-    let vk = groth16_solana::GROTH16_VK_BYTES;
+    let vk = sp1_solana::GROTH16_VK_BYTES;
 
     // Verify the proof.
     let result = verify_proof_fixture(&fixture, &vk);
