+++ example/sp1-program/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "fibonacci-program"
+version = "1.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+sp1-zkvm = { path = "../../../crates/zkvm/entrypoint" }
