+++ example/sp1-program/src/main.rs
@@ -1,0 +1,37 @@
+//! A simple program that takes a number `n` as input, and writes the `n-1`th and `n`th fibonacci
+//! number as an output.
+
+// These two lines are necessary for the program to properly compile.
+//
+// Under the hood, we wrap your main function with some extra code so that it behaves properly
+// inside the zkVM.
+#![no_main]
+sp1_zkvm::entrypoint!(main);
+
+pub fn main() {
+    // Read an input to the program.
+    //
+    // Behind the scenes, this compiles down to a system call which handles reading inputs
+    // from the prover.
+    let n = sp1_zkvm::io::read::<u32>();
+
+    // Write n to public input
+    sp1_zkvm::io::commit(&n);
+
+    // Compute the n'th fibonacci number, using normal Rust code.
+    let mut a = 0;
+    let mut b = 1;
+    for _ in 0..n {
+        let mut c = a + b;
+        c %= 7919; // Modulus to prevent overflow.
+        a = b;
+        b = c;
+    }
+
+    // Write the output of the program.
+    //
+    // Behind the scenes, this also compiles down to a system call which handles writing
+    // outputs to the prover.
+    sp1_zkvm::io::commit(&a);
+    sp1_zkvm::io::commit(&b);
+}
